requirements.txt
aiohttp
aiocoap
paho-mqtt
influxdb-client
